requirements.txt
Flask
twilio
pytesseract
python-dotenv
numpy
thefuzz==0.22.1
azure-cognitiveservices-vision-computervision>=0.9.0
requests==2.31.0
opencv-python==4.10.0.84
Pillow==10.4.0
pdf2image==1.17.0